SPO_backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.129.0
idna==3.11
psycopg2-binary
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn[standard]
minio
python-multipart
sqlalchemy
asyncpg
passlib
bcrypt==3.2.2
PyJWT
python-jose[cryptography]==3.3.0
dotenv
pydantic[email]
bs4
ebooklib